Removing deprecation
